moon: Modeling Obesity in Norway (MOON) Markov Cohort Model

Implements the MOON Markov cohort model for simulating obesity prevalence, incremental health-care costs, and years of life lost in a Norwegian birth cohort over the life course. Originally described in Bjørnelv et al. (2021) doi:10.1177/0272989X20971589 ; this package is an R port and refactor of the Excel/VBA implementation.

See also

Author

Maintainer: Nicolai Damslund nsimonsen@health.sdu.dk (ORCID)

Authors:

Nicolai Damslund nsimonsen@health.sdu.dk (ORCID)

Gudrun M W Bjørnelv (ORCID) (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))

Yufei Jiang yjiang@health.sdu.dk (ORCID) (Ported the MOON model from the original Excel/VBA implementation to R)

Magnus Skov Larsen maskla@health.sdu.dk (ORCID) (Ported the MOON model from the original Excel/VBA implementation to R)

Christina Hansen Edwards christina.h.edwards@ntnu.no (ORCID)

Vidar Halsteinli (ORCID) (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))

Rønnaug A Ødegaard (Original MOON model (Bjørnelv et al. 2021, doi:10.1177/0272989X20971589))

Line Planck Kongstad lpkongstad@health.sdu.dk (ORCID)